+++ stubdev/cuda_cholesky.h
@@ -4,6 +4,6 @@
 
 #include <eigen3/Eigen/Dense>
 
-double MatrixSolve(std::vector<Eigen::MatrixXd>& M,
-                   std::vector<Eigen::MatrixXd>& b,
-                   std::vector<Eigen::MatrixXd>& x);
+float MatrixSolve(std::vector<Eigen::MatrixXf>& M,
+                  std::vector<Eigen::MatrixXf>& b,
+                  std::vector<Eigen::MatrixXf>& x);
