+++ stubdev/cuda_gpu_collision.cuh
@@ -13,45 +13,45 @@
 
 // Define sphere geometry
 struct Sphere {
-  Eigen::Vector3d center;
-  double radius;
-  double mass;
+  Eigen::Vector3f center;
+  float radius;
+  float mass;
 
-  Eigen::Vector3d velocity;
+  Eigen::Vector3f velocity;
 
   // material properties
-  double stiffness;
-  double damping;
+  float stiffness;
+  float damping;
 };
 
 // Define a plane geometry
 struct Plane {
-  Eigen::Vector3d n;
-  Eigen::Vector3d p1;
-  Eigen::Vector3d p2;
+  Eigen::Vector3f n;
+  Eigen::Vector3f p1;
+  Eigen::Vector3f p2;
 
   // material properties
-  double stiffness;
-  double damping;
+  float stiffness;
+  float damping;
 };
 
 // Structure to hold collision data
 struct CollisionData {
   bool isColliding;
-  Eigen::Vector3d p_WC;       // Collision point on object A
-  Eigen::Vector3d nhat_BA_W;  // Collision normal, we follow the convention,
+  Eigen::Vector3f p_WC;       // Collision point on object A
+  Eigen::Vector3f nhat_BA_W;  // Collision normal, we follow the convention,
                               // pointing from B to A
 
-  double phi0;        // overlap distance
-  Eigen::Matrix3d R;  // rotation matrix
+  float phi0;         // overlap distance
+  Eigen::Matrix3f R;  // rotation matrix
 
-  double vn;  // normal relative velocity, positive if two spheres approaching,
-              // negative if separating; on the direction of contact normal
+  float vn;  // normal relative velocity, positive if two spheres approaching,
+             // negative if separating; on the direction of contact normal
 
   // variables for current step
 
   // history variables
-  double f_0;  // contact force history, f0 = k * x0
+  float f_0;  // contact force history, f0 = k * x0
 };
 
 // define a SAP data strucutre
@@ -81,45 +81,45 @@ struct CollisionGPUData {
     return d_collisionMatrixSpherePlane;
   }
 
-  __device__ const double* get_jacobian() const { return d_jacobian; }
+  __device__ const float* get_jacobian() const { return d_jacobian; }
 
-  __device__ double* get_jacobian() { return d_jacobian; }
+  __device__ float* get_jacobian() { return d_jacobian; }
 
   __device__ const int* get_num_collisions() const { return d_num_collisions; }
 
   __device__ int* get_num_collisions() { return d_num_collisions; }
 
-  __device__ const double* get_dynamic_matrix() const {
+  __device__ const float* get_dynamic_matrix() const {
     return d_dynamic_matrix;
   }
 
-  __device__ double* get_dynamic_matrix() { return d_dynamic_matrix; }
+  __device__ float* get_dynamic_matrix() { return d_dynamic_matrix; }
 
-  __device__ const double* get_velocity_vector() const {
+  __device__ const float* get_velocity_vector() const {
     return d_velocity_vector;
   }
 
-  __device__ double* get_velocity_vector() { return d_velocity_vector; }
+  __device__ float* get_velocity_vector() { return d_velocity_vector; }
 
-  __device__ const double* get_phi0() const { return d_phi0; }
+  __device__ const float* get_phi0() const { return d_phi0; }
 
-  __device__ double* get_phi0() { return d_phi0; }
+  __device__ float* get_phi0() { return d_phi0; }
 
-  __device__ const double* get_v_star() const { return d_v_star; }
+  __device__ const float* get_v_star() const { return d_v_star; }
 
-  __device__ double* get_v_star() { return d_v_star; }
+  __device__ float* get_v_star() { return d_v_star; }
 
-  __device__ const double* get_contact_stiffness() const {
+  __device__ const float* get_contact_stiffness() const {
     return d_contact_stiffness;
   }
 
-  __device__ double* get_contact_stiffness() { return d_contact_stiffness; }
+  __device__ float* get_contact_stiffness() { return d_contact_stiffness; }
 
-  __device__ const double* get_contact_damping() const {
+  __device__ const float* get_contact_damping() const {
     return d_contact_damping;
   }
 
-  __device__ double* get_contact_damping() { return d_contact_damping; }
+  __device__ float* get_contact_damping() { return d_contact_damping; }
 
   __device__ int get_num_problems() const { return num_problems; }
 
@@ -141,14 +141,14 @@ struct CollisionGPUData {
 
   // CPU data retrival, assuming the CPU data is already allocated
   void RetrieveCollisionDataToCPU(CollisionData* h_collisionMatrixSpheres);
-  void RetrieveJacobianToCPU(double* h_jacobian);
+  void RetrieveJacobianToCPU(float* h_jacobian);
   void RetrieveNumCollisionsToCPU(int* h_num_collisions);
-  void RetrieveDynamicMatrixToCPU(double* h_dynamic_matrix);
-  void RetrieveVelocityVectorToCPU(double* h_velocity_vector);
-  void RetrieveVStarToCPU(double* h_v_star);
-  void RetrievePhi0ToCPU(double* h_phi0);
-  void RetrieveContactStiffnessToCPU(double* h_contact_stiffness);
-  void RetrieveContactDampingToCPU(double* h_contact_damping);
+  void RetrieveDynamicMatrixToCPU(float* h_dynamic_matrix);
+  void RetrieveVelocityVectorToCPU(float* h_velocity_vector);
+  void RetrieveVStarToCPU(float* h_v_star);
+  void RetrievePhi0ToCPU(float* h_phi0);
+  void RetrieveContactStiffnessToCPU(float* h_contact_stiffness);
+  void RetrieveContactDampingToCPU(float* h_contact_damping);
   void RetieveSphereDataToCPU(Sphere* h_spheres);
 
   Sphere* GetSpherePtr() { return d_spheres; }
@@ -161,21 +161,21 @@ struct CollisionGPUData {
     return d_collisionMatrixSpherePlane;
   }
 
-  double* GetJacobianPtr() { return d_jacobian; }
+  float* GetJacobianPtr() { return d_jacobian; }
 
   int* GetNumCollisionsPtr() { return d_num_collisions; }
 
-  double* GetDynamicMatrixPtr() { return d_dynamic_matrix; }
+  float* GetDynamicMatrixPtr() { return d_dynamic_matrix; }
 
-  double* GetVelocityVectorPtr() { return d_velocity_vector; }
+  float* GetVelocityVectorPtr() { return d_velocity_vector; }
 
-  double* GetPhi0Ptr() { return d_phi0; }
+  float* GetPhi0Ptr() { return d_phi0; }
 
-  double* GetVStarPtr() { return d_v_star; }
+  float* GetVStarPtr() { return d_v_star; }
 
-  double* GetContactStiffnessPtr() { return d_contact_stiffness; }
+  float* GetContactStiffnessPtr() { return d_contact_stiffness; }
 
-  double* GetContactDampingPtr() { return d_contact_damping; }
+  float* GetContactDampingPtr() { return d_contact_damping; }
 
   CollisionGPUData* GetCollisionGPUDataPtr() { return d_collision_gpu_data; }
 
@@ -187,15 +187,15 @@ struct CollisionGPUData {
   Plane* d_planes;
   CollisionData* d_collisionMatrixSpheres;
   CollisionData* d_collisionMatrixSpherePlane;
-  double* d_jacobian;
+  float* d_jacobian;
   int* d_num_collisions;
-  double* d_dynamic_matrix;
-  double* d_velocity_vector;
-  double* d_phi0;
-  double* d_v_star;
+  float* d_dynamic_matrix;
+  float* d_velocity_vector;
+  float* d_phi0;
+  float* d_v_star;
 
-  double* d_contact_stiffness;
-  double* d_contact_damping;
+  float* d_contact_stiffness;
+  float* d_contact_damping;
 
   int num_problems = 0;
   int num_spheres = 0;
@@ -211,13 +211,12 @@ struct CollisionGPUData {
 __device__ CollisionData CheckSphereCollision(const Sphere& a,
                                               const Sphere& b) {
   CollisionData data = {
-      false, {0, 0, 0}, {0, 0, 0}, 0, Eigen::Matrix3d::Zero()};
+      false, {0, 0, 0}, {0, 0, 0}, 0, Eigen::Matrix3f::Zero()};
 
-  Eigen::Vector3d dist = a.center - b.center;
-  double distSquared =
-      dist(0) * dist(0) + dist(1) * dist(1) + dist(2) * dist(2);
-  double distLength = sqrt(distSquared);
-  double radiusSum = a.radius + b.radius;
+  Eigen::Vector3f dist = a.center - b.center;
+  float distSquared = dist(0) * dist(0) + dist(1) * dist(1) + dist(2) * dist(2);
+  float distLength = sqrt(distSquared);
+  float radiusSum = a.radius + b.radius;
 
   dist.normalize();
 
@@ -229,7 +228,7 @@ __device__ CollisionData CheckSphereCollision(const Sphere& a,
 
     data.nhat_BA_W.normalize();
     // Calculate collision points
-    Eigen::Vector3d midpoint;
+    Eigen::Vector3f midpoint;
     midpoint(0) = (a.center(0) + b.center(0)) / 2;
     midpoint(1) = (a.center(1) + b.center(1)) / 2;
     midpoint(2) = (a.center(2) + b.center(2)) / 2;
@@ -240,13 +239,13 @@ __device__ CollisionData CheckSphereCollision(const Sphere& a,
 
     // Get collision frame matrix
     // Random vector v is default to {1.0, 1.0, 1.0}
-    Eigen::Vector3d v(1.0, 1.0, 1.0);
+    Eigen::Vector3f v(1.0, 1.0, 1.0);
     v.normalize();
 
-    double y_hat_temp = v.dot(data.nhat_BA_W);
-    Eigen::Vector3d y_hat = v - y_hat_temp * data.nhat_BA_W;
+    float y_hat_temp = v.dot(data.nhat_BA_W);
+    Eigen::Vector3f y_hat = v - y_hat_temp * data.nhat_BA_W;
     y_hat.normalize();
-    Eigen::Vector3d x_hat = y_hat.cross(data.nhat_BA_W);
+    Eigen::Vector3f x_hat = y_hat.cross(data.nhat_BA_W);
 
     data.R(0, 0) = x_hat(0);           // x of x-axis
     data.R(0, 1) = x_hat(1);           // y of x-axis
@@ -289,9 +288,9 @@ __device__ void DetectSphereCollisions(const Sphere* spheres, int numProblems,
 __device__ CollisionData CheckSpherePlaneCollision(const Plane& p,
                                                    const Sphere& s) {
   CollisionData data = {
-      false, {0, 0, 0}, {0, 0, 0}, 0, Eigen::Matrix3d::Zero()};
+      false, {0, 0, 0}, {0, 0, 0}, 0, Eigen::Matrix3f::Zero()};
 
-  double distLength = (s.center - p.p1).dot(p.n);
+  float distLength = (s.center - p.p1).dot(p.n);
 
   if (distLength < s.radius) {
     data.isColliding = true;
@@ -303,13 +302,13 @@ __device__ CollisionData CheckSpherePlaneCollision(const Plane& p,
     // Get collision frame matrix
     // Random vector v is default to {1.0, 1.0, 1.0}
     // This matrix is normal-dependent
-    Eigen::Vector3d v(1.0, 1.0, 1.0);
+    Eigen::Vector3f v(1.0, 1.0, 1.0);
     v.normalize();
 
-    double y_hat_temp = v.dot(data.nhat_BA_W);
-    Eigen::Vector3d y_hat = v - y_hat_temp * data.nhat_BA_W;
+    float y_hat_temp = v.dot(data.nhat_BA_W);
+    Eigen::Vector3f y_hat = v - y_hat_temp * data.nhat_BA_W;
     y_hat.normalize();
-    Eigen::Vector3d x_hat = y_hat.cross(data.nhat_BA_W);
+    Eigen::Vector3f x_hat = y_hat.cross(data.nhat_BA_W);
 
     data.R(0, 0) = x_hat(0);           // x of x-axis
     data.R(0, 1) = x_hat(1);           // y of x-axis
@@ -352,19 +351,19 @@ __device__ void DetectSpherePlaneCollisions(const Sphere* spheres,
 __device__ void ConstructJacobianGamma(
     const Sphere* spheres, const Plane* planes, int numProblems, int numSpheres,
     int numPlanes, CollisionData* collisionMatrixSS,
-    CollisionData* collisionMatrixSP, double* d_jacobian, int* d_num_collisions,
-    double* d_phi0, double* d_contact_stiffness, double* d_contact_damping) {
+    CollisionData* collisionMatrixSP, float* d_jacobian, int* d_num_collisions,
+    float* d_phi0, float* d_contact_stiffness, float* d_contact_damping) {
   int idx = threadIdx.x;
   int p_idx = blockIdx.x;
 
-  Eigen::Map<Eigen::MatrixXd> full_jacobian(
+  Eigen::Map<Eigen::MatrixXf> full_jacobian(
       d_jacobian +
           blockIdx.x * (numSpheres * 3) * (numSpheres * numSpheres * 3),
       numSpheres * numSpheres * 3, numSpheres * 3);
-  Eigen::Map<Eigen::VectorXd> contact_stiffness(
+  Eigen::Map<Eigen::VectorXf> contact_stiffness(
       d_contact_stiffness + blockIdx.x * numSpheres * numSpheres,
       numSpheres * numSpheres, 1);
-  Eigen::Map<Eigen::VectorXd> contact_damping(
+  Eigen::Map<Eigen::VectorXf> contact_damping(
       d_contact_damping + blockIdx.x * numSpheres * numSpheres,
       numSpheres * numSpheres, 1);
 
@@ -394,12 +393,12 @@ __device__ void ConstructJacobianGamma(
             collisionMatrixSS[(p_idx * numSpheres * numSpheres) +
                               j * numSpheres + k]
                 .R *
-            Eigen::MatrixXd::Identity(3, 3);
+            Eigen::MatrixXf::Identity(3, 3);
         full_jacobian.block<3, 3>(collision_idx * 3, k * 3) =
             collisionMatrixSS[(p_idx * numSpheres * numSpheres) +
                               j * numSpheres + k]
                 .R *
-            -Eigen::MatrixXd::Identity(3, 3);
+            -Eigen::MatrixXf::Identity(3, 3);
 
         // add data to phi0
         d_phi0[p_idx * numSpheres * numSpheres + collision_idx] =
@@ -439,7 +438,7 @@ __device__ void ConstructJacobianGamma(
             collisionMatrixSP[(p_idx * numSpheres * numPlanes) +
                               k * numSpheres + j]
                 .R *
-            Eigen::MatrixXd::Identity(3, 3);
+            Eigen::MatrixXf::Identity(3, 3);
 
         // add data to phi0
         d_phi0[p_idx * numSpheres * numSpheres + collision_idx] =
@@ -453,23 +452,25 @@ __device__ void ConstructJacobianGamma(
   __syncwarp();
 }
 
-__device__ void ConstructDynamicMatrixVelocityVector(
-    const Sphere* spheres, int numProblems, int numSpheres,
-    double* d_dynamic_matrix, double* d_velocity_vector) {
+__device__ void ConstructDynamicMatrixVelocityVector(const Sphere* spheres,
+                                                     int numProblems,
+                                                     int numSpheres,
+                                                     float* d_dynamic_matrix,
+                                                     float* d_velocity_vector) {
   int idx = threadIdx.x;
   int p_idx = blockIdx.x;
 
-  Eigen::Map<Eigen::MatrixXd> dynamic_matrix(
+  Eigen::Map<Eigen::MatrixXf> dynamic_matrix(
       d_dynamic_matrix + blockIdx.x * numSpheres * 3 * numSpheres * 3,
       numSpheres * 3, numSpheres * 3);
-  Eigen::Map<Eigen::VectorXd> velocity_vector(
+  Eigen::Map<Eigen::VectorXf> velocity_vector(
       d_velocity_vector + blockIdx.x * numSpheres * 3, numSpheres * 3, 1);
 
   for (int j = idx; j < numSpheres; j += blockDim.x) {
     if (j < numSpheres) {
       dynamic_matrix.block<3, 3>(j * 3, j * 3) =
           spheres[p_idx * numSpheres + j].mass *
-          Eigen::MatrixXd::Identity(3, 3);
+          Eigen::MatrixXf::Identity(3, 3);
       velocity_vector.block<3, 1>(j * 3, 0) =
           spheres[p_idx * numSpheres + j].velocity;
     }
@@ -480,20 +481,20 @@ __device__ void ConstructDynamicMatrixVelocityVector(
 
 __device__ void CalculateFreeMotionVelocity(const Sphere* spheres,
                                             int numProblems, int numSpheres,
-                                            double* d_velocity_vector,
-                                            double* d_v_star) {
+                                            float* d_velocity_vector,
+                                            float* d_v_star) {
   int idx = threadIdx.x;
   int p_idx = blockIdx.x;
 
-  Eigen::Map<Eigen::VectorXd> velocity_vector(
+  Eigen::Map<Eigen::VectorXf> velocity_vector(
       d_velocity_vector + blockIdx.x * numSpheres * 3, numSpheres * 3, 1);
-  Eigen::Map<Eigen::VectorXd> v_star(d_v_star + blockIdx.x * numSpheres * 3,
+  Eigen::Map<Eigen::VectorXf> v_star(d_v_star + blockIdx.x * numSpheres * 3,
                                      numSpheres * 3, 1);
 
   for (int j = idx; j < numSpheres; j += blockDim.x) {
     if (j < numSpheres) {
       v_star.block<3, 1>(j * 3, 0) = velocity_vector.block<3, 1>(j * 3, 0) +
-                                     dt * Eigen::Vector3d(0, -4.905, 0);
+                                     dt * Eigen::Vector3f(0, -4.905, 0);
     }
   }
